requirements.txt
streamlit
torch
torchvision
pillow
numpy
nltk
matplotlib
cloudpickle